+++ test/IdentityServer4.Contrib.Caching.Redis.Tests/IdentityServerBuilderExtensionsIntegrationTests.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Net;
 using Elders.RedLock;
 using IdentityServer4.Contrib.Caching.Redis.Configuration;
 using IdentityServer4.Contrib.Caching.Redis.Tests.Misc;
@@ -7,6 +8,7 @@ using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Caching.StackExchangeRedis;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using StackExchange.Redis;
 using Xunit;
 
 namespace IdentityServer4.Contrib.Caching.Redis.Tests
@@ -50,6 +52,29 @@ namespace IdentityServer4.Contrib.Caching.Redis.Tests
             {
                 options.Configuration = this.configurationFixture.RedisCacheOptions.Configuration;
                 options.InstanceName = this.configurationFixture.RedisCacheOptions.InstanceName;
+            });
+
+            provider.GetRequiredService<IOptions<RedisCacheGrantStoreConfiguration>>();
+            provider.GetRequiredService<IOptions<RedisCacheOptions>>();
+            provider.GetRequiredService<IOptions<RedLockOptions>>();
+            provider.GetRequiredService<IDistributedCache>();
+            provider.GetRequiredService<IPersistedGrantStore>();
+            provider.GetRequiredService<IRedisLockManager>();
+        }
+        
+        [Fact]
+        public void IdentityServerBuilderExtensions_Register_Types_Options_Builder_For_Connection_Resolveable()
+        {
+            var provider = this.serviceProviderFixture.BuildDefaultServiceProvider(options =>
+            {
+                options.InstanceName = this.configurationFixture.RedisCacheOptions.InstanceName;
+                options.ConfigurationOptions = new ConfigurationOptions
+                {
+                    EndPoints =
+                    {
+                        new IPEndPoint(IPAddress.Parse("127.0.0.1"), 6379)
+                    }
+                };
             });
 
             provider.GetRequiredService<IOptions<RedisCacheGrantStoreConfiguration>>();